// src/main/resources/www2/src/api/api.ts
export interface Device {
    /**
     * 
     * @type {Set<DeviceAttribute>}
     * @memberof Device
     */
    'attributes'?: Set<DeviceAttribute>;
    /**
     * 
     * @type {boolean}
     * @memberof Device
     */
    'autoTryCredentials'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'changeDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'comments'?: string;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'contact'?: string;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'createdDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'creator'?: string;
    /**
     * 
     * @type {DeviceCredentialSet}
     * @memberof Device
     */
    'specificCredentialSet'?: DeviceCredentialSet;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'driver'?: string;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'eolDate'?: string;
    /**
     * 
     * @type {Module}
     * @memberof Device
     */
    'eolModule'?: Module;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'eosDate'?: string;
    /**
     * 
     * @type {Module}
     * @memberof Device
     */
    'eosModule'?: Module;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'family'?: string;
    /**
     * 
     * @type {number}
     * @memberof Device
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'location'?: string;
    /**
     * 
     * @type {Network4Address}
     * @memberof Device
     */
    'mgmtAddress'?: Network4Address;
    /**
     * 
     * @type {Domain}
     * @memberof Device
     */
    'mgmtDomain'?: Domain;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'networkClass'?: DeviceNetworkClassEnum;
    /**
     * 
     * @type {Set<DeviceGroup>}
     * @memberof Device
     */
    'ownerGroups'?: Set<DeviceGroup>;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'serialNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'softwareLevel'?: DeviceSoftwareLevelEnum;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'softwareVersion'?: string;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'status'?: DeviceStatusEnum;
    /**
     * 
     * @type {number}
     * @memberof Device
     */
    'sshPort'?: number;
    /**
     * 
     * @type {number}
     * @memberof Device
     */
    'telnetPort'?: number;
    /**
     * 
     * @type {Network4Address}
     * @memberof Device
     */
    'connectAddress'?: Network4Address;
    /**
     * 
     * @type {Array<number>}
     * @memberof Device
     */
    'credentialSetIds'?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof Device
     */
    'realDeviceType'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Device
     */
    'endOfLife'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Device
     */
    'endOfSale'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Device
     */
    'compliant'?: boolean;
}

export const DeviceNetworkClassEnum = {
    Firewall: 'FIREWALL',
    Loadbalancer: 'LOADBALANCER',
    Router: 'ROUTER',
    Server: 'SERVER',
    Switch: 'SWITCH',
    Switchrouter: 'SWITCHROUTER',
    Accesspoint: 'ACCESSPOINT',
    Wirelesscontroller: 'WIRELESSCONTROLLER',
    Consoleserver: 'CONSOLESERVER',
    Unknown: 'UNKNOWN'
} as const;

export type DeviceNetworkClassEnum = typeof DeviceNetworkClassEnum[keyof typeof DeviceNetworkClassEnum];
export const DeviceSoftwareLevelEnum = {
    Gold: 'GOLD',
    Silver: 'SILVER',
    Bronze: 'BRONZE',
    Unknown: 'UNKNOWN'
} as const;

export type DeviceSoftwareLevelEnum = typeof DeviceSoftwareLevelEnum[keyof typeof DeviceSoftwareLevelEnum];
export const DeviceStatusEnum = {
    Disabled: 'DISABLED',
    Inproduction: 'INPRODUCTION',
    Preproduction: 'PREPRODUCTION'
} as const;

export type DeviceStatusEnum = typeof DeviceStatusEnum[keyof typeof DeviceStatusEnum];

/**
 * 
 * @export
 * @interface DeviceAttribute
 */
export interface DeviceAttribute {
    /**
     * 
     * @type {string}
     * @memberof DeviceAttribute
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceAttribute
     */
    'type': string;
}
// ...
export interface DeviceCredentialSet {
    /**
     * 
     * @type {number}
     * @memberof DeviceCredentialSet
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof DeviceCredentialSet
     */
    'name'?: string;
    /**
     * 
     * @type {Domain}
     * @memberof DeviceCredentialSet
     */
    'mgmtDomain'?: Domain;
    /**
     * 
     * @type {boolean}
     * @memberof DeviceCredentialSet
     */
    'deviceSpecific'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof DeviceCredentialSet
     */
    'type': string;
}
// ...
export interface DeviceGroup {
    /**
     * 
     * @type {string}
     * @memberof DeviceGroup
     */
    'changeDate'?: string;
    /**
     * 
     * @type {number}
     * @memberof DeviceGroup
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof DeviceGroup
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceGroup
     */
    'folder'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof DeviceGroup
     */
    'hiddenFromReports'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof DeviceGroup
     */
    'type': string;
}
// ...
export interface DiscoverDeviceTypeTask extends TaskRestApiView {
    /**
     * 
     * @type {Network4Address}
     * @memberof DiscoverDeviceTypeTask
     */
    'deviceAddress'?: Network4Address;
    /**
     * 
     * @type {Device}
     * @memberof DiscoverDeviceTypeTask
     */
    'device'?: Device;
    /**
     * 
     * @type {number}
     * @memberof DiscoverDeviceTypeTask
     */
    'snapshotTaskId'?: number;
    /**
     * 
     * @type {string}
     * @memberof DiscoverDeviceTypeTask
     */
    'discoveredDeviceTypeDescription'?: string;
    /**
     * 
     * @type {number}
     * @memberof DiscoverDeviceTypeTask
     */
    'deviceId'?: number;
}
// ...
export interface Domain {
    /**
     * 
     * @type {number}
     * @memberof Domain
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof Domain
     */
    'changeDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof Domain
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Domain
     */
    'description'?: string;
    /**
     * 
     * @type {Network4Address}
     * @memberof Domain
     */
    'server4Address'?: Network4Address;
    /**
     * 
     * @type {Network6Address}
     * @memberof Domain
     */
    'server6Address'?: Network6Address;
}
// ...
export interface Module {
    /**
     * 
     * @type {number}
     * @memberof Module
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof Module
     */
    'slot'?: string;
    /**
     * 
     * @type {string}
     * @memberof Module
     */
    'partNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof Module
     */
    'serialNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof Module
     */
    'firstSeenDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof Module
     */
    'lastSeenDate'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Module
     */
    'removed'?: boolean;
}
// ...
export interface Network4Address {
    /**
     * 
     * @type {number}
     * @memberof Network4Address
     */
    'prefixLength'?: number;
    /**
     * 
     * @type {string}
     * @memberof Network4Address
     */
    'addressUsage'?: Network4AddressAddressUsageEnum;
    /**
     * 
     * @type {string}
     * @memberof Network4Address
     */
    'ip'?: string;
}

export const Network4AddressAddressUsageEnum = {
    Primary: 'PRIMARY',
    Secondary: 'SECONDARY',
    Vrrp: 'VRRP',
    Hsrp: 'HSRP',
    Secondaryvrrp: 'SECONDARYVRRP',
    Secondaryhsrp: 'SECONDARYHSRP'
} as const;

export type Network4AddressAddressUsageEnum = typeof Network4AddressAddressUsageEnum[keyof typeof Network4AddressAddressUsageEnum];
// ...
export interface TakeSnapshotTask extends TaskRestApiView {
    /**
     * 
     * @type {Device}
     * @memberof TakeSnapshotTask
     */
    'device'?: Device;
    /**
     * 
     * @type {boolean}
     * @memberof TakeSnapshotTask
     */
    'automatic'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof TakeSnapshotTask
     */
    'dontRunDiagnostics'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof TakeSnapshotTask
     */
    'dontCheckCompliance'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof TakeSnapshotTask
     */
    'deviceId'?: number;
}
// ...
export interface TaskRestApiView {
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'author'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'changeDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'comments'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'creationDate'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TaskRestApiView
     */
    'debugEnabled'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'executionDate'?: string;
    /**
     * 
     * @type {number}
     * @memberof TaskRestApiView
     */
    'id'?: number;
    /**
     * 
     * @type {TaskRestApiViewLogs}
     * @memberof TaskRestApiView
     */
    'logs'?: TaskRestApiViewLogs;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'scheduleReference'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'scheduleType'?: TaskRestApiViewScheduleTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof TaskRestApiView
     */
    'scheduleFactor'?: number;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'status'?: TaskRestApiViewStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'target'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'runnerId'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'nextExecutionDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'taskDescription'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TaskRestApiView
     */
    'repeating'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'type': string;
}

export const TaskRestApiViewScheduleTypeEnum = {
    Asap: 'ASAP',
    At: 'AT',
    Daily: 'DAILY',
    Monthly: 'MONTHLY',
    Weekly: 'WEEKLY',
    Hourly: 'HOURLY'
} as const;

export type TaskRestApiViewScheduleTypeEnum = typeof TaskRestApiViewScheduleTypeEnum[keyof typeof TaskRestApiViewScheduleTypeEnum];
export const TaskRestApiViewStatusEnum = {
    Cancelled: 'CANCELLED',
    Failure: 'FAILURE',
    New: 'NEW',
    Running: 'RUNNING',
    Scheduled: 'SCHEDULED',
    Success: 'SUCCESS',
    Waiting: 'WAITING'
} as const;

export type TaskRestApiViewStatusEnum = typeof TaskRestApiViewStatusEnum[keyof typeof TaskRestApiViewStatusEnum];

/**
 * 
 * @export
 * @interface TaskRestApiViewLogs
 */
export interface TaskRestApiViewLogs {
    /**
     * 
     * @type {number}
     * @memberof TaskRestApiViewLogs
     */
    'length'?: number;
}